requirements.txt
asgiref==3.2.5
Authlib==0.14.1
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-annoying==0.10.6
django-bootstrap4==1.1.1
django-extensions==2.2.8
django-filter==2.2.0
django-jsonstore==0.4.1
django-tables2==2.3.0
django-test-without-migrations==0.6
django-viewflow==1.6.0
django-widget-tweaks==1.4.8
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.9
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.8
